[Fix] Character Select
캐릭터 선택 버그 수정
