+++ Assets/1.Scene/Hur/CharacterSelected_hur.unity
@@ -38,7 +38,7 @@ RenderSettings:
   m_ReflectionIntensity: 1
   m_CustomReflection: {fileID: 0}
   m_Sun: {fileID: 0}
-  m_IndirectSpecularColor: {r: 0.44657874, g: 0.49641275, b: 0.5748172, a: 1}
+  m_IndirectSpecularColor: {r: 0.44657898, g: 0.4964133, b: 0.5748178, a: 1}
   m_UseRadianceAmbientProbe: 0
 --- !u!157 &3
 LightmapSettings:
@@ -453,8 +453,8 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 482839728}
-        m_TargetAssemblyTypeName: CharacterManager, Assembly-CSharp
+      - m_Target: {fileID: 482839727}
+        m_TargetAssemblyTypeName: CharacterSelect, Assembly-CSharp
         m_MethodName: ToggleLeft
         m_Mode: 1
         m_Arguments:
@@ -528,7 +528,6 @@ GameObject:
   - component: {fileID: 208404178}
   - component: {fileID: 208404177}
   - component: {fileID: 208404176}
-  - component: {fileID: 208404180}
   m_Layer: 0
   m_Name: Bird
   m_TagString: Untagged
@@ -566,8 +565,8 @@ Rigidbody:
   m_Mass: 1
   m_Drag: 0
   m_AngularDrag: 0.05
-  m_UseGravity: 1
-  m_IsKinematic: 0
+  m_UseGravity: 0
+  m_IsKinematic: 1
   m_Interpolate: 0
   m_Constraints: 0
   m_CollisionDetection: 0
@@ -635,18 +634,6 @@ LODGroup:
     renderers:
     - renderer: {fileID: 274648730}
   m_Enabled: 1
---- !u!114 &208404180
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 208404174}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: bafa0a16a911f3c4c8dd7d1f54dc36a0, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1 &222692931
 GameObject:
   m_ObjectHideFlags: 0
@@ -935,7 +922,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 482839726}
-  - component: {fileID: 482839728}
+  - component: {fileID: 482839727}
   m_Layer: 0
   m_Name: Character_List
   m_TagString: Untagged
@@ -951,7 +938,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 482839725}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 0.33830807, z: 0}
+  m_LocalPosition: {x: 0, y: 18.219309, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children:
   - {fileID: 208404175}
@@ -960,7 +947,7 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 3
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &482839728
+--- !u!114 &482839727
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -969,18 +956,13 @@ MonoBehaviour:
   m_GameObject: {fileID: 482839725}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: feef83ffafe2e444c9aaaac28ef1d638, type: 3}
+  m_Script: {fileID: 11500000, guid: 7f58c5107e3a8fa4c8accf3857491ed3, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   animals:
   - {fileID: 208404174}
   - {fileID: 7372094024700308233}
   - {fileID: 2605857209181307335}
-  curr_Animal: 0
-  ready: 0
-  bird: {fileID: 208404180}
-  fish: {fileID: 7372094024700308236}
-  monkey: {fileID: 2605857209181307338}
 --- !u!1 &631987099
 GameObject:
   m_ObjectHideFlags: 0
@@ -1132,8 +1114,8 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 482839728}
-        m_TargetAssemblyTypeName: CharacterManager, Assembly-CSharp
+      - m_Target: {fileID: 482839727}
+        m_TargetAssemblyTypeName: CharacterSelect, Assembly-CSharp
         m_MethodName: Selected_char
         m_Mode: 1
         m_Arguments:
@@ -1919,8 +1901,8 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 482839728}
-        m_TargetAssemblyTypeName: CharacterManager, Assembly-CSharp
+      - m_Target: {fileID: 482839727}
+        m_TargetAssemblyTypeName: CharacterSelect, Assembly-CSharp
         m_MethodName: ToggleRight
         m_Mode: 1
         m_Arguments:
@@ -2275,12 +2257,56 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1515205692}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 0.109, z: 0}
+  m_LocalPosition: {x: 0, y: 17.99, z: 0}
   m_LocalScale: {x: 1, y: 0.1923, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
   m_RootOrder: 2
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!1 &1518437135
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 1518437136}
+  - component: {fileID: 1518437137}
+  m_Layer: 0
+  m_Name: Character Manager
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &1518437136
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1518437135}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_RootOrder: 12
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &1518437137
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1518437135}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 35c3d90012a47b041b149e44325d73c3, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  curr_Animal: 0
 --- !u!1 &1565183556
 GameObject:
   m_ObjectHideFlags: 0
@@ -2728,7 +2754,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1692108573}
   m_LocalRotation: {x: 0.063097954, y: -0.92737126, z: 0.12861998, w: 0.34562716}
-  m_LocalPosition: {x: 1.249, y: 1.296, z: 1.499}
+  m_LocalPosition: {x: 1.249, y: 19.177, z: 1.499}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
@@ -3504,7 +3530,6 @@ GameObject:
   - component: {fileID: 8755597095595475207}
   - component: {fileID: 2605857209181307337}
   - component: {fileID: 2605857209181307336}
-  - component: {fileID: 2605857209181307338}
   m_Layer: 0
   m_Name: Monkey
   m_TagString: Untagged
@@ -3537,23 +3562,11 @@ Rigidbody:
   m_Mass: 1
   m_Drag: 0
   m_AngularDrag: 0.05
-  m_UseGravity: 1
-  m_IsKinematic: 0
+  m_UseGravity: 0
+  m_IsKinematic: 1
   m_Interpolate: 0
   m_Constraints: 0
   m_CollisionDetection: 0
---- !u!114 &2605857209181307338
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2605857209181307335}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: a08afa4ac28b6354381c8c4f57f00d3f, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1 &2615574036102961329
 GameObject:
   m_ObjectHideFlags: 0
@@ -4082,7 +4095,6 @@ GameObject:
   - component: {fileID: 4315869066918570953}
   - component: {fileID: 7372094024700308235}
   - component: {fileID: 7372094024700308234}
-  - component: {fileID: 7372094024700308236}
   m_Layer: 0
   m_Name: Fish
   m_TagString: Untagged
@@ -4115,23 +4127,11 @@ Rigidbody:
   m_Mass: 1
   m_Drag: 0
   m_AngularDrag: 0.05
-  m_UseGravity: 1
-  m_IsKinematic: 0
+  m_UseGravity: 0
+  m_IsKinematic: 1
   m_Interpolate: 0
   m_Constraints: 0
   m_CollisionDetection: 0
---- !u!114 &7372094024700308236
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7372094024700308233}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: f6f4dc65b77b3114daf7687dd1f68158, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1 &7462718316931915922
 GameObject:
   m_ObjectHideFlags: 0
