--- Assets/3.Script/Hur/Character_selelcted_hur/Ani_MonkeyControll.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a08afa4ac28b6354381c8c4f57f00d3f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
