--- Assets/3.Script/Hur/Character_selelcted_hur/CharacterManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: feef83ffafe2e444c9aaaac28ef1d638
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
