+++ Assets/3.Script/Jinwon/CharacterSelect.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 42f044b3f54f5354da2006d292ca0144
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
