+++ Assets/3.Script/Jinwon/CharacterSelect/CharacterManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f6f4dc65b77b3114daf7687dd1f68158
+guid: 35c3d90012a47b041b149e44325d73c3
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
