+++ Assets/3.Script/Jinwon/CharacterSelect/CharacterSelect.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: bafa0a16a911f3c4c8dd7d1f54dc36a0
+guid: 7f58c5107e3a8fa4c8accf3857491ed3
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
