+++ RankInfoFile/Score0.json
@@ -1,3 +1,3 @@
 {
-  "Score": 500
+  "Score": 600
 }
