+++ RankInfoFile/Score1.json
@@ -1,3 +1,3 @@
 {
-  "Score": 400
+  "Score": 600
 }
