+++ RankInfoFile/Score2.json
@@ -1,3 +1,3 @@
 {
-  "Score": 300
+  "Score": 500
 }
