+++ RankInfoFile/Score4.json
@@ -1,3 +1,3 @@
 {
-  "Score": 200
+  "Score": 400
 }
